+++ fiddles/ember/fiddle-0004-Mixins/app/app.js
@@ -2,20 +2,22 @@ import Ember from 'ember';
 import Resolver from './resolver';
 import loadInitializers from 'ember-load-initializers';
 import config from './config/environment';
+import numericBox from './numeric_box';
 
 let App;
 
 Ember.MODEL_FACTORY_INJECTIONS = true;
 
 App = Ember.Application.extend({
-  didReceiveAttrs() {
-    console.log('didReceiveAttrs')
-  },
   modulePrefix: config.modulePrefix,
   podModulePrefix: config.podModulePrefix,
   Resolver
 });
 
 loadInitializers(App, config.modulePrefix);
+
+var textbox = numericBox.create();
+console.log(textbox.valid("12"));//true
+textbox.onFocus();// do something creative on focus"
 
 export default App;