+++ fiddles/ember/fiddle-0004-Mixins/app/numeric_box.js
@@ -1,0 +1,7 @@
+import Ember from 'ember';
+import validNumberMixin from './mixins/valid-number';
+import focusableMixin from './mixins/focusable';
+
+export default Ember.Object.extend(validNumberMixin,focusableMixin,{
+  text:""
+});