+++ fiddles/node/fiddle-0014-Express/server-express.js
@@ -1,0 +1,14 @@
+#!/usr/bin/env node --harmony
+
+'use strict';
+const
+  express = require('express'),
+  app = express();
+app.use(express.logger('dev'));
+app.get('/api/:name', function (req, res) {
+  res.json(200, {"hello": req.params.name});
+});
+app.listen(3000, function () {
+  console.log("server-express.js started");
+  console.log("waiting for '/api/:name' request ...");
+});