+++ fiddles/node/fiddle-0014-Express/server.js
@@ -1,0 +1,12 @@
+#!/usr/bin/env node --harmony
+
+const
+  http = require('http'),
+  server = http.createServer(function (req, res) {
+    res.writeHead(200, {'Content-Type': 'text/plain'});
+    res.end('Hello World\n');
+  });
+
+server.listen(3000, function () {
+  console.log('ready captain!');
+});