+++ fiddles/node/fiddle-0014-Express/test-curl.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+curl -i http://localhost:3000/api/mom
+